bugfix/#124 fix leak for multiple fd in redirect struct
